Moved ''saveMessage'' method to imports/api/messaging
