+++ studious-octo-spork/.meteor/versions
@@ -47,6 +47,7 @@ less@2.5.7
 livedata@1.0.17
 localstorage@1.0.8
 logging@1.0.11
+mdg:validated-method@1.0.2
 meteor@1.1.13
 meteor-base@1.0.3
 minifier-css@1.1.10
