+++ studious-octo-spork/imports/startup/server/index.js
@@ -1,2 +1,3 @@
 import './fixtures.js';
-import './accounts-config.js'
+import './accounts-config.js';
+import './register-api.js';
