+++ studious-octo-spork/imports/startup/server/register-api.js
@@ -1,0 +1,1 @@
+import '../../api/messaging/methods.js'
